add useDrinksHook
