+++ src/components/DrinksSearch/DrinksSearch.jsx
@@ -1,9 +1,9 @@
 import { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useDispatch} from 'react-redux';
 import { useSearchParams } from 'react-router-dom';
 import { useParams } from 'react-router-dom/dist';
 import Select from 'react-select';
-// import debounce from 'lodash.debounce';
+import debounce from 'lodash.debounce';
 // components
 import Dots from 'components/Spinner/Dots';
 import { Paginator } from 'components/Paginator/Paginator';
@@ -12,8 +12,8 @@ import DrinkItemCard from './DrinkItemCard/DrinkItemCard';
 import { NoRecipe } from 'components/NoRecipe/NoRecipe';
 // redux
 import { fetchCategories, fetchDrinks, fetchIngredients } from 'redux/Drinks/DrinksOperation';
-import { getAllCategories, getAllIngredients, getDrinks, getIsLoading, getDrinksPage, getTotalPages } from 'redux/Drinks/DrinksSelectors';
 import { changeDrinksPage } from 'redux/Drinks/DrinksSlice';
+import useDrinks from 'hooks/useDrinks';
 // styles
 import { selectStylesCoktails, selectStylesIngredients } from './selectStyles';
 import css from './DrinksSearch.module.css';
@@ -23,18 +23,12 @@ export const DrinksSearch = () => {
   const { categoryName } = useParams();
   const [searchParams, setSearchParams] = useSearchParams({ category: categoryName });
 
-  const categoryList = useSelector(getAllCategories);
-  const ingredientList = useSelector(getAllIngredients);
-  const isLoading = useSelector(getIsLoading);
-  const entities = useSelector(getDrinks);
-  const currentPage = useSelector(getDrinksPage);
-  const totalPages = useSelector(getTotalPages);
+  const { categoryList, ingredientList, isLoading, entities, currentPage, totalPages } = useDrinks();
 
   const category = searchParams.get('category') || '';
   const ingredient = searchParams.get('ingredient') || '';
   const q = searchParams.get('q') || '';
   const page = searchParams.get('page') || 1;
-
 
   const categorySelectOptions = categoryList.map(({ name }) => {
     return { value: `${name}`, label: `${name}` };
@@ -53,21 +47,16 @@ export const DrinksSearch = () => {
     dispatch(fetchDrinks({ category, ingredient, q, page }));
   }, [dispatch, category, ingredient, q, page]);
 
-  // const debouncedHandleChange = debounce(payload => {
-  //   setQ(payload);
-  // }, 1000);
+  const debouncedHandleChange = e =>
+    debounce(e => {
+      setSearchParams({ q: e.target.value });
+    }, 1000);
 
-  
   return (
-    <>
+    <section>
       <form className={css.drinkRequestForm}>
         <label className={css.inputContainer}>
-          <input
-            id="inputSearch"
-            onChange={e => setSearchParams({ q: e.target.value })}
-            className={css.inputDrinks}
-            placeholder="Enter the text"
-          />
+          <input id="inputSearch" onChange={debouncedHandleChange} className={css.inputDrinks} placeholder="Enter the text" />
           {window.innerWidth > 768 && <SearchSvg className={css.searchSvg} />}
         </label>
 
@@ -99,14 +88,10 @@ export const DrinksSearch = () => {
             ))}
           </ul>
         )}
-
         {isLoading && <Dots className={css.loading} />}
-        {entities.length === 0 && isLoading === false && <NoRecipe title={'No results'} />}
-
-        {totalPages > 1 && (
-          <Paginator pages={{ page: currentPage, totalPages: totalPages }} onChangePage={changeDrinksPage} />
-        )}
+        {entities?.length === 0 && isLoading === false && <NoRecipe title={'No results'} />}
+        {totalPages > 1 && <Paginator pages={{ page: currentPage, totalPages: totalPages }} onChangePage={changeDrinksPage} />}
       </div>
-    </>
+    </section>
   );
 };
